fixed key input to work right, changed order of precedence on step so user cant avoid all battles by continually opening the menu

+++ BladeQuest/src/bladequest/system/BqThread.java
@@ -88,7 +88,7 @@ public class BqThread extends Thread
 				for(int ke : events)
 				{
 					onKeyEvent(ke);
-					Global.activity.keyEvents.remove(ke);
+					Global.activity.keyEvents.remove((Integer)ke);
 				}	
 			}			
     	}
